Fix compiler warnings about narrowing, name hiding, unused variables or parameters. Silence some msvc warnings.
(cherry picked from commit a63fc7a994e2d150251865bf08cb4f654cec65ee)

+++ include/boost/geometry/algorithms/detail/overlay/sort_by_side.hpp
@@ -303,7 +303,7 @@ struct side_sorter
             return;
         }
 
-        int const last = 1 + m_ranked_points.back().rank;
+        size_t const last = 1 + m_ranked_points.back().rank;
 
         // Move iterator after rank==0
         bool has_first = false;
